fixed delete logics

+++ controllers/delete_product.php
@@ -5,7 +5,27 @@ require_once("dbconn.php");
 if ($_SERVER['REQUEST_METHOD'] == 'GET') {
 
   $product_id = $_GET['id'];
-  echo $product_id;
+
+  $stmt = $conn->prepare("SELECT COUNT(*) FROM order_items WHERE product_id = ?");
+  $stmt->execute([$product_id]);
+  $orderItemCount = $stmt->fetchColumn();
+
+
+  $stmt = $conn->prepare("SELECT COUNT(*) FROM wishlist WHERE product_id = ?");
+  $stmt->execute([$product_id]);
+  $wishlistCount = $stmt->fetchColumn();
+
+  if ($orderItemCount > 0) {
+    $_SESSION['error'] = "Cannot delete: Product in order !";
+    header("Location: ../admin/admin.php?to=products");
+    exit();
+  }
+
+  if ($wishlistCount > 0) {
+    $query = "DELETE FROM wishlist where product_id =?;";
+    $stmt = $conn->prepare($query);
+    $stmt->execute([$product_id]);
+  }
 
   $sql = "delete from products where product_id = ?;";
   $stmt =   $conn->prepare($sql);
